Cambios hechos en la armonia del codigo y se añadio el hool _bag_ como carrito de compras
