+++ src/Components/Bag.jsx
@@ -1,0 +1,5 @@
+import HandleBag from "../HandleBag";
+
+function Bag(HandleBag){
+    console.log(HandleBag);
+}
