+++ src/Components/Catalog.jsx
@@ -1,5 +1,6 @@
 import { useState, useEffect } from "react";
 import typeColor from "../Color";
+import { HandleFetch } from "../HandleFetch";
 function Catalog(){
     const [paginaActual, setPaginaActual] = useState(1);
     const [filter, setFilter] = useState(1);
@@ -29,66 +30,22 @@ function Catalog(){
             setPaginaActual(1); 
         }
     }
+    function handleBag(card){
+        SetBag([...bag, card]);
+        console.log(bag);
+    }
+
     useEffect(() =>{
+
         const fetchData = async () => {
             try {
-                //delimitar url final
-                const url= `https://pokeapi.co/api/v2/${url2}`
-                const response = await fetch(url);
-                const jsonData = await response.json();
-                if(!isNaN(filter)){
-                    //filtramos pokemones por nombre, tipo, id, imagen
-                    const dataPromises = jsonData.pokemon_species.map((pokemon) => fetch(pokemon.url));
-                    const responses = await Promise.all(dataPromises);
-                    const jsonDataPromises = responses.map((response) => response.json());
-                    const cardData = await Promise.all(jsonDataPromises);
-                    
-                    const pokemonUrl = cardData.map((pokemon) => fetch(pokemon.varieties[0].pokemon.url));
-                    const Pokeresponse= await Promise.all(pokemonUrl);
-                    const jsonDatas= Pokeresponse.map((resp) => resp.json());
-                    const data= await Promise.all(jsonDatas);
-                    
-                    const card= data.map((info) => {
-                        return {
-                            id: info.id,
-                            name: info.name,
-                            image: `https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/${info.id}.png`,
-                            types: info.types.map((type) => type.type.name),
-                            stats: {
-                                attack: info.stats[1].base_stat,
-                                defense: info.stats[2].base_stat,
-                                speed: info.stats[5].base_stat
-                            }
-                        }
-                    })
-        
-                    setCarddata(card);
-                }
-                else{
-                    const dataPromises = jsonData.pokemon.map((pokemon) => fetch(pokemon.pokemon.url));
-                    const responses = await Promise.all(dataPromises);
-                    const jsonDataPromises = responses.map((response) => response.json());
-                    const data = await Promise.all(jsonDataPromises);
-
-                    const card= data.map((info) => {
-                        return {
-                            id: info.id,
-                            name: info.name,
-                            image: `https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/${info.id}.png`,
-                            types: info.types.map((type) => type.type.name),
-                            stats: {
-                                attack: info.stats[1].base_stat,
-                                defense: info.stats[2].base_stat,
-                                speed: info.stats[5].base_stat
-                            } 
-                        }
-                    })
-                    setCarddata(card);
-                }
+                const fetchedData = await HandleFetch(url2, filter); 
+                setCarddata(fetchedData.card); 
             } catch (error) {
-                console.error("Error fetching data:", error);
+                console.error('Error fetching data:', error);
             }
-        }
+        };
+
         fetchData();
     }, [filter]);
     return(
@@ -125,8 +82,7 @@ function Catalog(){
                                 <p>Speed</p>
                             </div>
                         </div>
-                        <button id="btn" onClick={()=> {SetBag([...bag, card])}}>ADD</button>
-                        {console.log(bag)}
+                        <button id="btn" onClick={()=> handleBag(card)}>ADD</button>
                     </div>
                 ))}
             </div>
